add websocket metrics
