+++ Sources/ParseSwift/LiveQuery/Protocols/LiveQueryable.swift
@@ -1,0 +1,12 @@
+//
+//  LiveQueryable.swift
+//  ParseSwift
+//
+//  Created by Corey Baker on 1/2/21.
+//  Copyright © 2021 Parse Community. All rights reserved.
+//
+
+import Foundation
+
+protocol LiveQueryable: Encodable {}
+
